rename coeffs files
Just rename coeffs files without functional change.
