+++ Makefile.am
@@ -27,7 +27,7 @@ CLEANFILES = src/src_sinc.s
 # MinGW requires -no-undefined if a DLL is to be built.
 src_libsamplerate_la_LDFLAGS = -no-undefined -version-info $(SHARED_VERSION_INFO) $(SHLIB_VERSION_ARG)
 src_libsamplerate_la_SOURCES = src/samplerate.c src/src_sinc.c src/src_zoh.c src/src_linear.c \
-	src/common.h src/fastest_coeffs.h src/mid_qual_coeffs.h src/high_qual_coeffs.h \
+	src/common.h src/coeffs_fastest.h src/coeffs_mid_qual.h src/coeffs_high_qual.h \
 	src/src_config.h
 
 #-------------------------------------------------------------------------------
