+++ src/coeffs_mid_qual.h
